fix: make retry attempts respect current slot number (#108)
# fix: make retry attempts respect current slot number 

Retry attempts were happening but if the slot changed in the middle of a retry attempt, it wouldn't recognize it and would get into an error loop, exhausting all retry attempts on an obviously incorrect block.

Instead, this PR checks the initial slot number and each retry attempt will only happen if the slot is still valid.

Closes ENG-1127