fix(sim): use previous block number for creating alloy db in simulator (#119)
# fix(sim): use previous block number for creating alloy db in simulator

PR #116 introduced a logic change that reused the block number from the `block_env` to fetch the state for the alloy DB instance, which results in querying for state of a block number that doesn't exist.

This commit reduces that number by 1 to correctly create the AlloyDB state from the previous block instead of the in-progress block.