fix: glibc runtime errors (#159)
# fix: glibc runtime errors

This PR updates the docker stage 0 to use a known version of `bookworm`​ in for the next `cargo-chef`​ build step instead of the `lukemathwalker/cargo-check`​ image that it previously used in order to maintain `glibc`​ compatibility with the `debian:bookwork-slim`​ image used in stage 3.

## Testing

Built a container with github actions and ran it in DevNet - tl;dr it landed blocks and the runtime error was gone.

Closes [ENG-1424](https://linear.app/initiates/issue/ENG-1424/builder-main-branch-builds-are-broken)