fix(flashbots): Do not block on submission (#184)
There's no point of blocking when we send the bundle to flashbots. The only thing we do after is record metrics. This will matter more on a production environment where network lag could occur on services we don't control (flashbots). I think blocking on quincey/getting the nonce on `prepare` is OK.